+++ Gruntfile.js
@@ -75,6 +75,21 @@ module.exports = function(grunt) {
 					template: 'test/templates/template.css'
 				}
 			},
+			template_scss: {
+				src: 'test/src/*.svg',
+				dest: 'test/tmp/template_scss',
+				options: {
+					stylesheet: 'scss',
+					template: 'test/templates/template.scss'
+				}
+			},
+			template_sass: {
+				src: 'test/src/*.svg',
+				dest: 'test/tmp/template_sass',
+				options: {
+					template: 'test/templates/template.sass'
+				}
+			},
 			html_template: {
 				src: 'test/src/*.svg',
 				dest: 'test/tmp/html_template',
@@ -146,8 +161,9 @@ module.exports = function(grunt) {
 					hashes: false
 				}
 			},
+			// #167: Ligatures with hypen don’t work
 			ligatures: {
-				src: 'test/src/*.svg',
+				src: 'test/src_ligatures/*.svg',
 				dest: 'test/tmp/ligatures',
 				options: {
 					hashes: false,
@@ -237,7 +253,22 @@ module.exports = function(grunt) {
 				options: {
 					hashes: false
 				}
-			}
+			},
+			woff2: {
+				src: 'test/src/*.svg',
+				dest: 'test/tmp/woff2',
+				options: {
+					types: 'woff2,woff'
+				}
+			},
+			woff2_node: {
+				src: 'test/src/*.svg',
+				dest: 'test/tmp/woff2_node',
+				options: {
+					types: 'woff2,woff',
+					engine: 'node'
+				}
+			},
 		},
 		nodeunit: {
 			all: ['test/webfont_test.js']
@@ -245,15 +276,7 @@ module.exports = function(grunt) {
 		jshint: {
 			all: ['Gruntfile.js', 'tasks/*.js', 'test/*.js'],
 			options: {
-				node: true,
-				white: false,
-				smarttabs: true,
-				eqeqeq: true,
-				immed: true,
-				latedef: false,
-				newcap: true,
-				undef: true,
-				laxbreak: true
+				jshintrc: true
 			}
 		},
 		watch: {